Applied 03-alg-tweaks.patch, from Eran Tromer.
Add earlyAbort and immediateUpdown.

1. Added an optional "earlyAbort" mode:
Stop checking alternatives as soon as you found a feasible
choice and the remaining ones have a lower priority.
Enabled by default.

NOTE: 
This is superceeded by the forking search patches later in this series,
and will be eliminated in a later patch. It's kept here as a lightweight
alternative to forking the search (and because redoing all the patches
inbetween is hard).

2. Added an optional "immediateUpdown" mode in which we avoid long and
useless remove cascades by doing _updown() immediately instead of
_remove()ing and postponing the updown to _pending. To make this work,
I modified _updown() to let you force package replacement/removal
(i.e., leaving as-is is not a permissible alternative).
Enabled by default.

See more details in the source comments.