Applied 05-prune.patch, from Eran Tromer.
Add pruneByWeight algorithm tweak.

This adds a mode that remembers the best feasible changeset
weight seen so far (in an appropriate sense), and prunes the search when
things becomes irreversibly worse than that. See the comment for details.

Alas, we have to cheat a bit about the "irreversibly" part, but I'm yet
to find a case where it matters.

This is a very effective heuristic when we guessed the order of the
alternatives correctly. The forking search in later patches eliminates
this luck factor.