Applied 07-pri-upgrade.patch, from Eran Tromer.
Make priority magnitudes affect PolicyUpgrade weights

This makes package weights matter in magnitude, not only in relation.
So with this mode enabled, it's possible to tell smart "always track
package foo on channel bar, no matter the cost" using priorities.

It can be disabled by prioritiesAffectWeight=0, which reverts
to a behavior similar (though not perfectly identical) to the old one.

The priorityScale variable should probably be made configurable in
runtime via smart's config mechanism.

The priorities-to-weight transformationn is placed in a seperate function
since we'll need to call it in another context in a future patch.