Applied 11-prune-by-prohibits.patch, from Eran Tromer.
Added lookahead and pruning in _install()

This extends the pruning patch by doing a static best-case look-ahead,
in order to quickly rule out installing packages that conflict with a
lot of (or high-priority) installed packages. It works as follows:

Whenever a package pkg is _install()ed, we compute the list of installed
packages which can *never* coexist with pkg, and optimistically compute
the minimal weight damage that can result from changing all those
"prohibited" packages. If even that best-case estimate is worse than the
best feasible weight we've already seen, we prune the search. Most of
the computation is cached.

This lets us quickly rule out those nasty attempted glibc downgrades
before trying to remove half the packases in the system.