Applied 14-task-class.patch, from Eran Tromer.
Wrap tasks in a Task class

Another no-op (hopefully!), but prepares the infrastructure for
forking search. We wrap each generator function in a Task class
with some useful attributes. This fulfill several purposes:

To send new values of pruneweight to the generator functions,
the caller need to share some mutable storage, and objects serve
nicey.

It lets us factor out some code for computing changeset weights
(only when necessary) and recording them.

It lets us implement a comparison operator which (in a future 
patch) will replace the "weight += keeporder" kludge.

The Task classes are created as nested classes inside
class Transaction, since they belong there logically (and
this avoids ruining the diff). 

Tasks receive two weight thresholds from their creator: a prune weight
at which to give by raising Prune, and a yield weight at which to
yield execution (but possibly resume later). In principle we could do
with just a yield weight, since prune is essentially the same as
yield-and-never-be-resumed; but explicitly raising Prune makes things
clearer and avoids zombie tasks taking up resources.

(The identiation is slightly weird to avoid ruining the diff;
this can be fixed once the code is finalized.)