Applied 16-yield-search.patch, from Eran Tromer.
Add a few yield points to the transaction search.

This adds a few yield points to the transaction generator
functions, to allow the search to switch to other alternatives.
The same caveats apply as for pruning: we're cheating a bit.

It would be nice to find more yield and prune points.